be/requirements.txt
fastapi
pillow
pyyaml
uvicorn
sqlalchemy
pytest
pytest-asyncio